refactor(config): removed old config
