+++ libs/config/common/src/lib/app-name.ts
@@ -1,0 +1,9 @@
+export type appName =
+  | 'platformApi'
+  | 'accountApi'
+  | 'combineApi'
+  | 'simulatorsApi'
+  | 'platform'
+  | 'account'
+  | 'dispatch'
+  | 'simulators';
