+++ libs/config/common/src/lib/app-routes.ts
@@ -1,4 +1,4 @@
-import { environment } from '@biosimulations/shared/environments';
+import { environment, envs } from '@biosimulations/shared/environments';
 import { EndpointLoader, LoadedEndpoints } from './endpointLoader';
 
 export class AppRoutes {
@@ -6,7 +6,7 @@ export class AppRoutes {
   private dispatchApp: string;
   private platformApp: string;
 
-  public constructor(env?: 'local' | 'dev' | 'stage' | 'prod') {
+  public constructor(env?: envs) {
     // We can read the env that is provided in the shared environment file as the default
     if (env == undefined) {
       env = environment.env;
