+++ libs/config/common/src/lib/config-common.ts
@@ -1,5 +1,5 @@
 export * from './ui';
-export * from './urls';
+export * from './app-name';
 export * from './endpoints';
 export * from './app-routes';
 export * from './resource-identifiers';
