+++ libs/config/nest/src/lib/biosimulations-server-config.ts
@@ -1,9 +1,10 @@
 import { registerAs } from '@nestjs/config';
 import { appName } from '@biosimulations/config/common';
+import { envs } from '@biosimulations/shared/environments';
 
 export default registerAs('server', () => {
   // What is the environment?
-  const env = process.env.ENV || 'dev';
+  const env: envs = (process.env.ENV as envs) || 'dev';
 
   // The name of the running app
   const app: appName = (process.env.APP as appName) || 'platformApi';
