+++ libs/shared/environments/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/environment';
+export * from './lib/environment.type';
